move Provider base trait to goose-providers
