+++ crates/goose-providers/src/goose_mode.rs
