+++ crates/goose-providers/src/lib.rs
@@ -3,9 +3,12 @@ pub mod canonical;
 pub mod conversation;
 pub mod errors;
 pub mod formats;
+pub mod goose_mode;
 pub mod images;
 pub mod json;
 pub(crate) mod mcp_utils;
 pub mod model;
+pub mod permission;
+pub mod retry;
 pub mod thinking;
 pub mod utils;
