+++ crates/goose-providers/src/permission.rs
