+++ crates/goose/src/config/mod.rs
@@ -2,7 +2,6 @@ pub mod base;
 pub mod declarative_providers;
 mod experiments;
 pub mod extensions;
-pub mod goose_mode;
 mod migrations;
 pub mod paths;
 pub mod permission;
@@ -21,7 +20,7 @@ pub use extensions::{
     get_extension_by_name, get_warnings, is_extension_enabled, remove_extension,
     resolve_extensions_for_new_session, set_extension, set_extension_enabled, ExtensionEntry,
 };
-pub use goose_mode::GooseMode;
+pub use goose_providers::goose_mode::GooseMode;
 pub use permission::PermissionManager;
 pub use signup_nanogpt::configure_nanogpt;
 pub use signup_openrouter::configure_openrouter;
