+++ crates/goose/src/permission/mod.rs
@@ -1,8 +1,10 @@
-pub mod permission_confirmation;
 pub mod permission_inspector;
 pub mod permission_judge;
 pub mod permission_store;
 
-pub use permission_confirmation::{Permission, PermissionConfirmation};
+pub use goose_providers::permission::{Permission, PermissionConfirmation};
+pub mod permission_confirmation {
+    pub use goose_providers::permission::PrincipalType;
+}
 pub use permission_inspector::PermissionInspector;
 pub use permission_store::ToolPermissionStore;
